+++ changelog.md
@@ -266,6 +266,7 @@ Breaking changes:
 
 ## Next
 
+* `create` throws error if item is its own parent
 * Throw specific error if try to create child of non-existent parent (closes #21)
 * Support sequelize's `fields` option
 * Refactor